% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-color-themes.R
\name{custom_palette}
\alias{custom_palette}
\title{Custom color palettes for \code{ggplot2}}
\usage{
custom_palette(
  palette = c("high_contrast", "redblue", "viridis"),
  viridis_option = "plasma"
)
}
\arguments{
\item{palette}{Character. The color palette to use. Options are:
\code{"high_contrast"} for a high-contrast color palette,
\code{"redblue"} for a red-blue color palette, and
\code{"viridis"} for a viridis color palette.}

\item{viridis_option}{Character. The viridis color palette to use. See
\code{\link[viridis]{viridis}} for available options.}
}
\description{
Custom color palettes for \code{ggplot2} scales.
For some inspiration, see https://github.com/Yu-Group/vthemes/blob/main/R/ggplot-scale-themes.R
}
\examples{
library(scales)
show_col(custom_palette()(8))
show_col(custom_palette(palette = "redblue")(5))
show_col(custom_palette(palette = "viridis")(8))

}
