% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-color-themes.R
\name{scale_color_vmodern}
\alias{scale_color_vmodern}
\alias{scale_colour_vmodern}
\alias{scale_fill_vmodern}
\title{Custom color scales}
\usage{
scale_colour_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)

scale_color_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)

scale_fill_vmodern(
  discrete = FALSE,
  palette = "high_contrast",
  viridis_option = "plasma",
  ...
)
}
\arguments{
\item{discrete}{Logical. If \code{TRUE}, use a discrete color scale. If
\code{FALSE}, use a continuous color scale.}

\item{palette}{Character. The color palette to use. Options are:
\code{"high_contrast"} for a high-contrast color palette,
\code{"redblue"} for a red-blue color palette, and
\code{"viridis"} for a viridis color palette.}

\item{viridis_option}{Character. The viridis color palette to use. See
\code{\link[viridis]{viridis}} for available options.}

\item{...}{Additional arguments to pass to \code{ggplot2::discrete_scale()}
if \code{discrete == TRUE} and to \code{viridis::scale_color_viridis()} or
\code{viridis::scale_fill_viridis()} if \code{discrete == FALSE}.}
}
\description{
Custom scale functions (fill and colour/color) for \code{ggplot2}.
For \code{discrete == FALSE}, the vmodern scale uses the \code{viridis} color
palette. For \code{discrete == TRUE}, custom color palettes are used.
}
\examples{
library(ggplot2)
ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Petal.Length) +
  geom_point() +
  scale_color_vmodern(discrete = FALSE)
ggplot(iris) +
  aes(x = Sepal.Length, fill = Species) +
  geom_density() +
  scale_fill_vmodern(discrete = TRUE)

}
