% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot2 theme}
\usage{
theme_custom()
}
\description{
A custom ggplot2 theme. For some inspiration,
see https://github.com/Yu-Group/vthemes/blob/main/R/ggplot-themes.R
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
 geom_point() +
 theme_custom()

}
